Código organizado em diferentes diretórios:screens, components e models
