+++ lib/screens/transferencia/formulario.dart
@@ -1,0 +1,58 @@
+import 'package:flutter/material.dart';
+
+import '../../components/editor.dart';
+import '../../models/transferencia.dart';
+
+class FormularioTransferencia extends StatefulWidget {
+
+  @override
+  State<StatefulWidget> createState() {
+    // TODO: implement createState
+    return FormularioTransferenciaState();
+  }
+}
+
+class FormularioTransferenciaState extends State<FormularioTransferencia> {
+
+  final TextEditingController _controladorCampoNumeroConta =
+  TextEditingController();
+  final TextEditingController _controladorCampoValor =
+  TextEditingController();
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('Criando trasferência'),),
+      body: SingleChildScrollView(
+        child: Column(
+          children: <Widget>[
+            Editor(
+                controlador: _controladorCampoNumeroConta,
+                rotulo: 'Número da conta',
+                dica: '0000-00'),
+            Editor(
+                controlador: _controladorCampoValor,
+                rotulo: 'Valor',
+                dica: '0.00',
+                icone: Icons.monetization_on),
+
+            ElevatedButton(
+              child: const Text('Confirmar'),
+              onPressed: () => _criaTransferencia(context),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+  void _criaTransferencia(BuildContext context) {
+    final int? numeroConta = int.tryParse(_controladorCampoNumeroConta.text);
+    final double? valor = double.tryParse(_controladorCampoValor.text);
+    if(numeroConta != null && valor != null){
+      final transferenciaCriada =  Transferencia(valor, numeroConta);
+      debugPrint('Criando Transferência');
+      debugPrint('$transferenciaCriada');
+      Navigator.pop(context, transferenciaCriada);
+    }
+  }
+}
