+++ lib/screens/transferencia/lista.dart
@@ -1,0 +1,66 @@
+import 'package:flutter/material.dart';
+import '../../models/transferencia.dart';
+import 'formulario.dart';
+
+class ListaTransferencias extends StatefulWidget {
+  final List<Transferencia?> _transferencias = [];
+  @override
+  State<StatefulWidget> createState() {
+    // TODO: implement createState
+    return ListaTransferenciasState();
+  }
+}
+
+class ListaTransferenciasState extends State<ListaTransferencias> {
+  @override
+  Widget build(BuildContext context) {
+    // TODO: implement build
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Transferências'),
+      ),
+      body: ListView.builder(
+        itemCount: widget._transferencias.length,
+        itemBuilder: (context, indice) {
+          final transferencia = widget._transferencias[indice];
+          return ItemTransferencia(transferencia!);
+        },
+      ),
+      floatingActionButton: FloatingActionButton(
+        child: const Icon(Icons.add),
+        onPressed: () {
+          final Future<Transferencia?> future = Navigator.push(context, MaterialPageRoute(builder: (context) {
+            return FormularioTransferencia();
+          }));
+          future.then((transferenviaRecebida) {
+            debugPrint('chegou no then future');
+            debugPrint('$transferenviaRecebida');
+            if (transferenviaRecebida != null) {
+              setState(() {
+                widget._transferencias.add(transferenviaRecebida);
+              });}
+          });
+        },
+      ),
+    );
+  }
+}
+
+class ItemTransferencia extends StatelessWidget {
+
+  final Transferencia _transferencia;
+
+  const ItemTransferencia(this._transferencia);
+
+  @override
+  Widget build(BuildContext context) {
+    // TODO: implement build
+    return Card(
+      child: ListTile(
+        leading: const Icon(Icons.monetization_on),
+        title: Text(_transferencia.valor.toString()),
+        subtitle: Text(_transferencia.numeroConta.toString()),
+      ),
+    );
+  }
+}
